requirement.txt
streamlit
pandas
gemini
python-dotenv
langchain
langchain-community
tiktoken